Add files via upload
This project contains the following files:
DatasetReduced.csv : This is the train dataset with reduced features which is used in MainR.m file.
ReleasedDatasetReduced.csv : This is the test dataset with reduced features which is used in MainR.m file.
ImagePreprocessing.py : Python file for image preprocessing. It is used as an input for RandomProjection.m file.
RandomProjection.py : Python file to reduce the features of original dataset. It is used as an input for knn.m file.
knn.m : K Nearest Neighbour, programmed in Matlab.